headless JSON エラー契約を補強
