+++ cmd/root.go
@@ -85,11 +85,21 @@ Examples:
 
 			resolvedOutputFormat, err := resolveOutputFormat(outputFormat, headless)
 			if err != nil {
+				if headless {
+					result := app.NewHeadlessUsageErrorResult("", "", err.Error()).
+						WithInput(newHeadlessPreRunInputMetadata(cmd, args))
+					return writeHeadlessResult(cmd, result, resolvedExitPolicy)
+				}
 				return commandErrorForExitPolicy(err, resolvedExitPolicy, 2)
 			}
 
 			mode, err := resolveExecutionMode(args, resolvedOutputFormat)
 			if err != nil {
+				if resolvedOutputFormat == outputFormatJSON {
+					result := app.NewHeadlessUsageErrorResult("", "", err.Error()).
+						WithInput(newHeadlessPreRunInputMetadata(cmd, args))
+					return writeHeadlessResult(cmd, result, resolvedExitPolicy)
+				}
 				return commandErrorForExitPolicy(err, resolvedExitPolicy, 2)
 			}
 			if headlessPromptFileFlagChanged(cmd) && mode != executionModeHeadless {
@@ -113,8 +123,9 @@ Examples:
 
 			switch mode {
 			case executionModeHeadless:
-				promptInput, err := resolveHeadlessPromptInput(cmd, args)
+				promptInput, err := resolveHeadlessPromptInput(cmd, args, imageFlag != "")
 				if err != nil {
+					promptInput.input = withHeadlessImageInputMetadataForCurrentProvider(promptInput.input, imageFlag)
 					result := app.NewHeadlessUsageErrorResult("", "", err.Error()).WithInput(promptInput.input)
 					return writeHeadlessResult(cmd, result, resolvedExitPolicy)
 				}
@@ -131,7 +142,16 @@ Examples:
 							WithInput(promptInput.input)
 						return writeHeadlessResult(cmd, result, resolvedExitPolicy)
 					}
-					return err
+					var configErr *headlessRuntimeSelectionConfigError
+					if errors.As(err, &configErr) {
+						promptInput.input = withHeadlessImageInputMetadataForProviderName(promptInput.input, imageFlag, configErr.provider)
+						result := app.NewHeadlessConfigErrorResult(configErr.provider, configErr.model, configErr.message).
+							WithInput(promptInput.input)
+						return writeHeadlessResult(cmd, result, resolvedExitPolicy)
+					}
+					promptInput.input = withHeadlessImageInputMetadataForCurrentProvider(promptInput.input, imageFlag)
+					result := app.NewHeadlessConfigErrorResult("", "", err.Error()).WithInput(promptInput.input)
+					return writeHeadlessResult(cmd, result, resolvedExitPolicy)
 				}
 				options := app.HeadlessRunOptions{
 					FailOnToolError: failOnToolError,
@@ -143,6 +163,9 @@ Examples:
 					return writeHeadlessResult(cmd, imageResolution.result, resolvedExitPolicy)
 				}
 				options.Image = imageResolution.image
+				if strings.TrimSpace(promptInput.query) == "" && options.Image != nil {
+					promptInput.query = app.DefaultImagePrompt
+				}
 				result := runHeadless(cmd.Context(), promptInput.query, runtime.model, runtime.provider, runtime.cfg, options)
 				if result == nil {
 					result = app.NewHeadlessConfigErrorResult(runtime.provider.Name(), runtime.model, "headless run returned nil result")
@@ -208,6 +231,27 @@ func (e *headlessProviderSetupRequiredError) Error() string {
 	return e.message
 }
 
+type headlessRuntimeSelectionConfigError struct {
+	provider string
+	model    string
+	message  string
+}
+
+func (e *headlessRuntimeSelectionConfigError) Error() string {
+	return e.message
+}
+
+func newHeadlessRuntimeSelectionConfigError(provider string, model string, err error) error {
+	if err == nil {
+		return nil
+	}
+	return &headlessRuntimeSelectionConfigError{
+		provider: provider,
+		model:    model,
+		message:  err.Error(),
+	}
+}
+
 type commandExitCodeError struct {
 	message string
 	code    int
@@ -248,6 +292,9 @@ func resolveProviderForExecutionMode(cmd *cobra.Command, providerName string, mo
 			model:    model,
 			message:  err.Error(),
 		}
+	}
+	if mode == executionModeHeadless {
+		return nil, newHeadlessRuntimeSelectionConfigError(providerName, model, err)
 	}
 	return nil, err
 }
