+++ internal/agent/image_prompt.go
@@ -1,0 +1,4 @@
+package agent
+
+// DefaultImagePrompt は画像だけが渡された時の既定プロンプト。
+const DefaultImagePrompt = "Please analyze this image."
